# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations": {
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created successfully"
          },
          "403": {
            "description": "User must be verified"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new organization (requires verified user)",
        "tags": [
          "Organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns list of organizations"
          }
        },
        "summary": "List all organizations",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/my": {
      "get": {
        "operationId": "OrganizationsController_findMyOrganizations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get organizations created by current user",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/pending-verifications": {
      "get": {
        "operationId": "OrganizationsController_getPendingVerifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns list of organizations pending verification"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Moderator] Get all organizations with pending verification",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns organization details"
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "summary": "Get organization by ID",
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated successfully"
          },
          "403": {
            "description": "Not authorized to update this organization"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update organization (creator only)",
        "tags": [
          "Organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization deleted successfully"
          },
          "403": {
            "description": "Not authorized to delete this organization"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete organization (soft delete, creator only)",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/{id}/request-verification": {
      "post": {
        "operationId": "OrganizationsController_requestVerification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification request submitted successfully"
          },
          "400": {
            "description": "Legal documents are required or verification already pending/completed"
          },
          "403": {
            "description": "Not authorized - only organization owner can request verification"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request verification for organization (requires legal documents to be uploaded first)",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/{id}/verification-status": {
      "patch": {
        "operationId": "OrganizationsController_updateVerificationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification status updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Moderator] Approve or reject organization verification",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/organizations/{id}/audit": {
      "get": {
        "operationId": "OrganizationsController_getAuditTrail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns blockchain audit history"
          }
        },
        "summary": "Get organization blockchain audit trail",
        "tags": [
          "Organizations"
        ]
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "operationId": "AdminController_getDashboardStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns dashboard statistics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get admin dashboard statistics",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns list of all users"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users (Admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminController_updateUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User role updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user role (Admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/audit/organizations/{id}": {
      "get": {
        "operationId": "AdminController_getOrganizationAudit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns blockchain audit trail"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get organization audit trail",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/audit/campaigns/{id}": {
      "get": {
        "operationId": "AdminController_getCampaignAudit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns blockchain audit trail"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get campaign audit trail",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/blockchain/summary": {
      "get": {
        "operationId": "AdminController_getBlockchainSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns blockchain stats"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get blockchain summary",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Transparent Donation System API",
    "description": "\n## Overview\nA blockchain-enabled donation platform ensuring transparency, auditability, and immutability.\n\n## Features\n- **User Management**: Registration, authentication, and verification\n- **Organization Management**: CRUD with blockchain recording\n- **Campaign Management**: Create and manage fundraising campaigns\n- **Donations**: Anonymous donation support with blockchain verification\n- **Verification System**: Multi-level verification for users, organizations, and campaigns\n- **Audit Trail**: Complete blockchain-backed audit history\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Roles\n- **USER**: Standard user account\n- **MODERATOR**: Reviews and approves verification requests\n- **AUDITOR**: Read-only access to verification and audit data\n- **ADMIN**: System administration and operational exception handling\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "description": "User ID who owns this organization (will be set from JWT token automatically)",
            "example": "676a1b2c3d4e5f6a7b8c9d0e",
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "Organization name",
            "example": "Quỹ Từ Thiện Trái Tim Nhân Ái"
          },
          "description": {
            "type": "string",
            "description": "Organization description (mission, vision, activities)",
            "example": "Tổ chức từ thiện phi lợi nhuận hoạt động trong lĩnh vực giáo dục và y tế cho trẻ em nghèo vùng cao"
          },
          "website": {
            "type": "string",
            "description": "Organization website URL",
            "example": "https://traitim-charity.org"
          },
          "contactEmail": {
            "type": "string",
            "description": "Contact email address",
            "example": "contact@traitim-charity.org"
          },
          "contactPhone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "0901234567"
          },
          "address": {
            "type": "string",
            "description": "Physical address / headquarters location",
            "example": "123 Lê Lợi, Quận 1, TP. Hồ Chí Minh"
          },
          "legalDocuments": {
            "description": "Legal documents URLs for verification (business license, registration certificate, etc.)",
            "example": [
              "https://storage.example.com/docs/license.pdf",
              "https://storage.example.com/docs/certificate.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "name"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name",
            "example": "Quỹ Từ Thiện Trái Tim Nhân Ái - Cập nhật"
          },
          "description": {
            "type": "string",
            "description": "Organization description (mission, vision, activities)",
            "example": "Tổ chức từ thiện phi lợi nhuận với mục tiêu mang lại cuộc sống tốt đẹp hơn cho trẻ em nghèo"
          },
          "website": {
            "type": "string",
            "description": "Organization website URL",
            "example": "https://traitim-charity.org"
          },
          "contactEmail": {
            "type": "string",
            "description": "Contact email address",
            "example": "info@traitim-charity.org"
          },
          "contactPhone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "0287654321"
          },
          "address": {
            "type": "string",
            "description": "Physical address / headquarters location",
            "example": "456 Nguyễn Huệ, Quận 1, TP. Hồ Chí Minh"
          },
          "legalDocuments": {
            "description": "Legal documents URLs for verification (business license, registration certificate, etc.)",
            "example": [
              "https://storage.example.com/docs/license.pdf",
              "https://storage.example.com/docs/certificate.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}